backend/requirements.txt
aiofiles==0.8.0
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2==2.9.3
python-dotenv==0.21.0
pytz==2022.2.1
sqlparse==0.4.2
